Implement HA store that is backed by a raft-replicated transaction log
The HA store uses raft to convert any store.Store instance into a
fault-tolerant store that is replicated across multiple nodes.

The store is comprised of two components:
- the raft subsystem which replicates a log of serialized []txn.Op
values and drives a FSM that applies them to the backing store.
- the cluster API service which allows nodes that are not yet members of
the raft cluster to discover the leader (via a simple gossip protocol)
and contact it so the leader can add them to the cluster. In addition,
the same API is used to relay transactions from followers to the leader
and relay back the result once the leader attempts to apply the
transaction.